add random pass fn

+++ mac/.hammerspoon/init.lua
@@ -45,7 +45,7 @@ function getLaunchFocusOrHideAndSwitchBackFn(bundleid, kill)
   kill = kill or false
   return function()
     currentApp = hs.application.frontmostApplication()
-    if currentApp and (currentApp:bundleID() == bundleid) then
+    if currentApp and (currentApp:bundleID() == bundleid) and not currentApp:isHidden() then
       if kill then
         currentApp:kill()
       else
